Add files via upload
This project is about Sockect Programming using Python. Sockets 
help you to establish these connections from server to client, and 
Python certainly simplifies connections. 
Sockets are the foundation of the networks.TCP sockets are used 
here to connect the server and client.
 So, we are designing a Music Player System where a particular 
category of music is selected in the server side and sent to client 
side.
The Client Side is responsible for controlling the music files via 
respective buttons available. Many will be able to listen the music at 
the same time which is achieved by multithreading.